Move sanitize_filename function to storage module
